Rename CalculatedFieldSupport to FieldGetterSupport (#4620)
The module hosts both calculated-field getters and source-field (`_src`) getters - the old name was specific to the first.
